Added functionality to Party
Added check for if party is full at a 3 cap. If it is then the monster does not get added. I have also adjusted the updateMonsterStatus function in accountBalance.js so that it searches for a menagerie first then searches for the monster in the menagerie, rather than searching the entire monster table.